Update button ordering option for accessibility
Since jQuery sortable is not keyboard accessible, I'm using number
inputs as an optional way to reorder buttons using the keyboard. Still
using sortable as well, although the two are difficult to integrate.

Initial load is always ordered according to settings both visually and
by number inputs. If the number inputs are changed, the visual order is
no longer correct, but dragging by hand is disabled.

There may be issues with buttons added, not sure if the nubmer inputs
will update properly; needs to be checked.