Refactor buttons classes
This is in preparation to set up a query_args filter but I had to move
everything around first because it was the same thing over and over
again. Attributes have had to be updated but I'm not sure that's a bad
thing--need to check on this more. At least now (almost) all Pinterest
functions are in the Pinterest class.