Update link decoding/encoding
Link parameters were being encoded and decoded in multiple places, which
was causing some output to be overencoded, and some to be missed (like
Yoast twitter titles). This removes all of the encoding to after the
query_args have been retrieved and just before they are added.

The button url is still escaped on final output.