Change plugin enqueuing
Considering this due to multiple tickets regarding manually placed
buttons without loaded styles--this way, the plugin style, if enabled,
will always be _registered_, but not necessarily _enqueued_. If the
plugin knows the buttons will be output, the style will be enqueued.
If the buttons are output and the style has been registered, but not
enqueued, the style will be loaded (in the footer, which does cause
FOUC).

This is related to commit a5d31b330b2994e1f5be2bc42308190e37c826bd,
which fixed the styles always loading on singular posts/pages/etc, but
allowed buttons to look correct with incorrect settings.